WIP: Rethink scraping design
After some frustrating attempts at scraping oldtown.se, I concluded that
my initial thoughts about scraping and updating is not very good.
It needs to be way more flexible and dynamic.
The current approach works well with sites such as lindholmen.se, where
the html is consistent, the menus are updated every day, and all info
about all restaurants can be found on the same site.
A lot of other restaurants though, are not specifically lunch
restaurants, and have a menu that rarely changes. Also many have
websites with inconsistent html, making them hard to scrape correctly.

So, we need some way to decide which sites should be scraped and updated
automatically, and even within those, which restaurants should be
updated manually or automatically. Might work with tags in the structs.

As a side note, many restaurants might be easiest to scrap from foodora,
so implementing a generic foodora scraper, could be worth the time,
since it could probably be reused a lot, since we then have consistent
html to scrape.