Version 0.1.0
Should be working about the same as the original Golang version.
Only a parser for lindholmen.se, and presentation in HTML and JSON.
Content beside Lindholmen can be added manually to the DB, and it will
show up.
Later versions will have the scraping logic reworked. This version is
just to be able to replace the old Golang version at parity level.